Moved recyclerViewWithCats to catsListFragment
