+++ app/src/main/java/com/example/catslist/viewmodels/CatsListFragmentViewModel.kt
@@ -1,7 +1,32 @@
 package com.example.catslist.viewmodels
 
+import android.app.DownloadManager
+import android.content.Context
+import android.net.Uri
+import android.os.Environment
+import android.util.Log
+import android.widget.Toast
 import androidx.lifecycle.ViewModel
+import com.example.catslist.tools.CatStorage
 
 class CatsListFragmentViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
+
+    fun addCat() {
+        Log.v("MainActivityViewModel","addCat")
+        CatStorage.getNewCatData()
+    }
+
+    fun downloadCatImage(context: Context, url: String, catId: String) {
+        val request = DownloadManager.Request(Uri.parse(url))
+        request.setTitle("Image of cat with id $catId")
+        request.setDescription("Downloading image..")
+        request.setDestinationInExternalPublicDir(Environment.DIRECTORY_DOWNLOADS, "cat_$catId.jpg")
+        request.setAllowedNetworkTypes(DownloadManager.Request.NETWORK_WIFI or DownloadManager.Request.NETWORK_MOBILE)
+        request.setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE_NOTIFY_COMPLETED)
+        request.allowScanningByMediaScanner()
+        val manager = context.getSystemService(Context.DOWNLOAD_SERVICE) as DownloadManager
+        manager.enqueue(request)
+        Toast.makeText(context, "Downloading started", Toast.LENGTH_LONG).show()
+    }
+
 }
