+++ app/src/main/java/com/example/catslist/viewmodels/MainActivityViewModel.kt
@@ -11,22 +11,5 @@ import com.example.catslist.tools.CatStorage
 
 class MainActivityViewModel: ViewModel() {
 
-    fun addCat() {
-        Log.v("MainActivityViewModel","addCat")
-        CatStorage.getNewCatData()
-    }
-
-    fun downloadCatImage(context: Context, url: String, catId: String) {
-        val request = DownloadManager.Request(Uri.parse(url))
-        request.setTitle("Image of cat with id $catId")
-        request.setDescription("Downloading image..")
-        request.setDestinationInExternalPublicDir(Environment.DIRECTORY_DOWNLOADS, "cat_$catId.jpg")
-        request.setAllowedNetworkTypes(DownloadManager.Request.NETWORK_WIFI or DownloadManager.Request.NETWORK_MOBILE)
-        request.setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE_NOTIFY_COMPLETED)
-        request.allowScanningByMediaScanner()
-        val manager = context.getSystemService(Context.DOWNLOAD_SERVICE) as DownloadManager
-        manager.enqueue(request)
-        Toast.makeText(context, "Downloading started", Toast.LENGTH_LONG).show()
-    }
 
 }
