+++ app/src/main/java/com/example/catslist/views/MainActivity.kt
@@ -14,47 +14,26 @@ import com.example.catslist.tools.CatStorage
 import com.example.catslist.tools.CatsListener
 import com.example.catslist.adapters.CatsActionsListener
 import com.example.catslist.adapters.CatsAdapter
+import com.example.catslist.adapters.ViewPageAdapter
 import com.example.catslist.room.*
 import com.example.catslist.viewmodels.MainActivityViewModel
+import com.google.android.material.tabs.TabLayoutMediator
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
 class MainActivity : AppCompatActivity() {
 
     private val tag  = "MainActivity"
-    private lateinit var adapter: CatsAdapter
     private lateinit var binding: ActivityMainBinding
-
-    private val catsStorage: CatStorage
-        get() = (applicationContext as App).catsService
+    private val fragmentsList = listOf(
+        CatsListFragment.newInstance(),
+        FavoriteCatsListFragment.newInstance()
+    )
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
-
-        adapter = CatsAdapter(object: CatsActionsListener{
-            override fun onAddToFavorites(cat: Cat) {
-                TODO("Not yet implemented")
-            }
-
-            override fun onDownload(cat: Cat) {
-                Log.v(tag, "onDownload()")
-                MainActivityViewModel().downloadCatImage(applicationContext, cat.url, cat.id)
-            }
-
-        })
-        val layoutManager = LinearLayoutManager(this)
-        binding.recyclerViewWithCats.layoutManager = layoutManager
-        binding.recyclerViewWithCats.adapter = adapter
-        binding.recyclerViewWithCats.addOnScrollListener(object : RecyclerView.OnScrollListener() {
-            override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
-                super.onScrolled(recyclerView, dx, dy)
-                if(!recyclerView.canScrollVertically(1)) {
-                    addCat()
-                }
-            }
-        })
 
         //testing room database
         lateinit var catsDao: CatsDao
@@ -86,14 +65,5 @@ class MainActivity : AppCompatActivity() {
         }.attach()
 
 
-    }
-
-    private val catsListener: CatsListener = {
-        Log.v("catsListener", "catsListener!")
-        adapter.catsList = it
-    }
-
-    private fun addCat() {
-        repeat(5) { MainActivityViewModel().addCat() }
     }
 }
